requirements.txt
python-dotenv
numpy
pandas
pymongo[srv]
certifi
scikit-learn
dill
PyYAML
# -e .
